blacklist form

+++ blacklisted/views.py
@@ -1,31 +1,46 @@
-from django.http import JsonResponse
+from django.shortcuts import render
+from django.http import HttpResponseRedirect, JsonResponse
+from django.shortcuts import get_object_or_404
+
+from django.contrib.auth.decorators import login_required
 
 from .models import BlackListedStudent
+from .forms import BlackListStudentForm
 from attendance.models import Entry
+from profiles.models import Student
 
 
-def blacklist_student(request):
-    '''Blacklist Student'''
+@login_required
+def blacklist_add(request, entry_id, std_id, att_id):
+    template_name = 'blacklisted/new_blacklist.html'
 
     if request.method == 'GET':
-        entry_id = request.GET.get('entry_id')
+        student = get_object_or_404(Student, id=std_id)
+        form = BlackListStudentForm(initial={'student': student,
+                                             'tmp_name': student})
+        field = form.fields['student']
+        field.widget = field.hidden_widget()
 
-        # Blacklist Selected Student
-        student_obj = None
-        try:
+        return render(request, template_name, {'form': form})
+
+    if request.method == 'POST':
+        student = get_object_or_404(Student, id=std_id)
+
+        form = BlackListStudentForm(request.POST, initial={'student': student,
+                                                           'tmp_name': student})
+        field = form.fields['student']
+        field.widget = field.hidden_widget()
+
+        if form.is_valid():
+            form.save()
+
+            # Delete entry from attendance entry
             entry_obj = Entry.objects.get(id=entry_id)
-            student_obj = entry_obj.student  # Save student object
             entry_obj.delete()
-            result = "success"
-        except Entry.DoesNotExist:
-            result = "error"
 
-        # Add Student in Blacklist List
-        if result == "success":
-            obj = BlackListedStudent(student=student_obj)
-            obj.save()
+            return HttpResponseRedirect('/attendances/' + str(att_id) + "/")
 
-    return JsonResponse({'result': result})
+        return render(request, template_name, {'form': form})
 
 
 def remove_blacklist(request):
